+++ src/TreeTools-init.c
@@ -15,7 +15,6 @@ extern SEXP _TreeTools_keep_tip(SEXP, SEXP);
 extern SEXP _TreeTools_minimum_spanning_tree(SEXP);
 extern SEXP _TreeTools_mixed_base_to_parent(SEXP, SEXP);
 extern SEXP _TreeTools_num_to_parent(SEXP, SEXP);
-extern SEXP _TreeTools_postorder_edges(SEXP);
 extern SEXP _TreeTools_postorder_order(SEXP);
 extern SEXP _TreeTools_preorder_edges_and_nodes(SEXP, SEXP);
 extern SEXP _TreeTools_preorder_weighted(SEXP, SEXP, SEXP);
@@ -44,7 +43,6 @@ static const R_CallMethodDef callMethods[] = {
   {"_TreeTools_minimum_spanning_tree", (DL_FUNC) &_TreeTools_minimum_spanning_tree, 1},
   {"_TreeTools_mixed_base_to_parent", (DL_FUNC) &_TreeTools_mixed_base_to_parent, 2},
   {"_TreeTools_num_to_parent", (DL_FUNC) &_TreeTools_num_to_parent, 2},
-  {"_TreeTools_postorder_edges", (DL_FUNC) &_TreeTools_postorder_edges, 1},
   {"_TreeTools_postorder_order", (DL_FUNC) &_TreeTools_postorder_order, 1},
   {"_TreeTools_preorder_edges_and_nodes", (DL_FUNC) &_TreeTools_preorder_edges_and_nodes, 2},
   {"_TreeTools_preorder_weighted", (DL_FUNC) &_TreeTools_preorder_weighted, 3},