+++ NEWS.md
@@ -15,8 +15,7 @@
 - Improve speed of `DropTip()`, by an order of magnitude in some cases.
 
 - Support edge lengths in `Preorder()`, `RootTree()`, `UnrootTree()` and
-  (provisionally) `Postorder()`
-  ([#49](https://github.com/ms609/TreeTools/issues/49),
+  `Postorder()` ([#49](https://github.com/ms609/TreeTools/issues/49),
   [#89](https://github.com/ms609/TreeTools/issues/89)).
 
 - Fix bug when tree is rooted on a discontinuous outgroup.