+++ NEWS.md
@@ -1,4 +1,4 @@
-# TreeTools 1.6.0.9008 #
+# TreeTools 1.6.0.9009 #
 
 - `as.MixedBase()` uniquely represents binary trees as a mixed-base vector.